requirements.txt
flask
flask_bcrypt
authlib
google-generativeai
pickle